Reuse object converters in migration (#5773)
* move converters to core level

* use object_converters in migration

* handle mixed non-pk objects in converter
