+++ Package.swift
@@ -80,6 +80,7 @@ let notSyncServerSources: [String] = [
     "realm/node.cpp",
     "realm/obj.cpp",
     "realm/obj_list.cpp",
+    "realm/object_converter.cpp",
     "realm/object_id.cpp",
     "realm/query.cpp",
     "realm/query_engine.cpp",
