+++ src/realm/CMakeLists.txt
@@ -45,6 +45,7 @@ set(REALM_SOURCES
     node.cpp
     mixed.cpp
     obj.cpp
+    object_converter.cpp
     global_key.cpp
     query_engine.cpp
     query_expression.cpp
@@ -169,6 +170,7 @@ set(REALM_INSTALL_HEADERS
     null.hpp
     obj.hpp
     obj_list.hpp
+    object_converter.hpp
     object_id.hpp
     owned_data.hpp
     query.hpp
